Add protein example from phylosim vignette
phylosim is an R-based sequence simulator. Unfortunately, it was written
by a Java programmer who doesn't know anything about R; it is a little
slow and awkward and makes heavy use of global variables are mutable
data structures. The disdain they have for the language they chose to
use is a bit irritating. But, I got the example working, but need to
review the mathematics behind what it is doing.